requirement.txt
asgiref                           ==  3.7.2
async-timeout                       ==   4.0.3
attrs                        ==          23.1.0
autobahn                          ==     23.6.2
Automat                 ==               22.10.0
cachetools                   ==          5.3.2
certifi                           ==     2023.7.22
cffi                 ==                  1.16.0
channels                  ==             4.0.0
channels-redis                 ==        4.1.0
charset-normalizer     ==                3.3.2
constantly                  ==           23.10.4
cryptography                     ==      41.0.5
daphne               ==                  4.0.0
defusedxml                ==             0.8.0rc2
Deprecated                     ==        1.2.14
Django                              ==   4.2.6
django-channels-jwt-auth-middleware   == 1.0.0
django-cors-headers                   == 4.3.0
django-oauth-toolkit        ==           2.3.0
djangorestframework              ==      3.14.0
djangorestframework-simplejwt         == 5.3.0
drf-social-oauth2                 ==     2.1.3
ecdsa                   ==               0.18.0
google-auth                  ==          2.24.0
hyperlink                         ==     21.0.0
idna                        ==           3.4
incremental                      ==      22.10.0
jwcrypto                              == 1.5.0
msgpack                     ==           1.0.7
mysqlclient                      ==      2.2.0
oauthlib                ==               3.2.2
Pillow                       ==          10.1.0
pip                               ==     23.3.1
pyasn1     ==                            0.5.0
pyasn1-modules  ==                       0.3.0
pycparser            ==                  2.21
PyJWT                     ==             2.8.0
pyOpenSSL                      ==        23.3.0
python-decouple                     ==   3.8
python-jose           ==                 3.3.0
python3-openid             ==            3.2.0
pytz                            ==       2023.3.post1
redis                                ==  5.0.1
requests        ==                       2.31.0
requests-oauthlib    ==                  1.3.1
rsa                       ==             4.9
service-identity               ==        23.1.0
setuptools                          ==   68.2.2
six                           ==         1.16.0
social-auth-app-django             ==    5.4.0
social-auth-core         ==              4.5.0
sqlparse                      ==         0.4.4
Twisted                            ==    23.10.0

txaio                       ==           23.1.1
typing_extensions                ==      4.8.0
tzdata                  ==               2023.3
urllib3                      ==          2.1.0
uuid                              ==     1.30
wheel       ==                           0.41.2
wrapt            ==                      1.16.0
zope.interface        ==                 6.1




